+++ stacker/providers/aws/interactive.py
@@ -38,6 +38,11 @@ def requires_replacement(changeset):
             'True']
 
 
+def get_raw_input(message):
+    """ Just a wrapper for raw_input for testing purposes. """
+    return raw_input(message)
+
+
 def ask_for_approval(full_changeset=None, include_verbose=False):
     """Prompt the user for approval to execute a change set.
 
@@ -52,7 +57,7 @@ def ask_for_approval(full_changeset=None, include_verbose=False):
     if include_verbose:
         approval_options.append('v')
 
-    approve = raw_input("Execute the above changes? [{}] ".format(
+    approve = get_raw_input("Execute the above changes? [{}] ".format(
         '/'.join(approval_options)))
 
     if include_verbose and approve == "v":
@@ -103,75 +108,136 @@ def output_summary(fqn, action, changeset, replacements_only=False):
     logger.info('%s %s:\n%s', fqn, action, summary)
 
 
-class Provider(AWSProvider):
-    """AWS Cloudformation Change Set Provider"""
+def wait_till_change_set_complete(cfn_client, change_set_id, try_count=25,
+                                  sleep_time=.5, max_sleep=3):
+    """ Checks state of a changeset, returning when it is in a complete state.
 
-    def __init__(self, *args, **kwargs):
-        self.replacements_only = kwargs.pop('replacements_only', False)
-        super(Provider, self).__init__(*args, **kwargs)
+    Since changesets can take a little bit of time to get into a complete
+    state, we need to poll it until it does so. This will try to get the
+    state `try_count` times, waiting `sleep_time` * 2 seconds between each try
+    up to the `max_sleep` number of seconds. If, after that time, the changeset
+    is not in a complete state it fails. These default settings will wait a
+    little over one minute.
 
-    def _wait_till_change_set_complete(self, change_set_id):
-        complete = False
-        response = None
-        while not complete:
-            response = retry_on_throttling(
-                self.cloudformation.describe_change_set,
-                kwargs={
-                    'ChangeSetName': change_set_id,
-                },
-            )
-            complete = response["Status"] in ("FAILED", "CREATE_COMPLETE")
-            if not complete:
-                time.sleep(2)
-        return response
+    Args:
+        cfn_client (:class:`botocore.client.CloudFormation`): Used to query
+            cloudformation.
+        change_set_id (str): The unique changeset id to wait for.
+        try_count (int): Number of times to try the call.
+        sleep_time (int): Time to sleep between attempts.
+        max_sleep (int): Max time to sleep during backoff
 
-    def update_stack(self, fqn, template_url, parameters, tags, **kwargs):
-        logger.debug("Attempting to create change set for stack: %s.", fqn)
+    Return:
+        dict: The response from cloudformation for the describe_change_set
+            call.
+    """
+    complete = False
+    response = None
+    for i in range(try_count):
         response = retry_on_throttling(
-            self.cloudformation.create_change_set,
-            kwargs={
-                'StackName': fqn,
-                'TemplateURL': template_url,
-                'Parameters': parameters,
-                'Tags': tags,
-                'Capabilities': ["CAPABILITY_NAMED_IAM"],
-                'ChangeSetName': get_change_set_name(),
-            },
-        )
-        change_set_id = response["Id"]
-        response = self._wait_till_change_set_complete(change_set_id)
-        if response["Status"] == "FAILED":
-            if "didn't contain changes" in response["StatusReason"]:
-                logger.debug(
-                    "Stack %s did not change, not updating.",
-                    fqn,
-                )
-                raise exceptions.StackDidNotChange
-            raise Exception(
-                "Failed to describe change set: {}".format(response)
-            )
-
-        if response["ExecutionStatus"] != "AVAILABLE":
-            raise Exception("Unable to execute change set: "
-                            "{}".format(response))
-
-        action = "replacements" if self.replacements_only else "changes"
-        changeset = response["Changes"]
-        if self.replacements_only:
-            changeset = requires_replacement(changeset)
-
-        if len(changeset):
-            output_summary(fqn, action, changeset,
-                           replacements_only=self.replacements_only)
-            ask_for_approval(
-                full_changeset=response["Changes"],
-                include_verbose=True,
-            )
-
-        retry_on_throttling(
-            self.cloudformation.execute_change_set,
+            cfn_client.describe_change_set,
             kwargs={
                 'ChangeSetName': change_set_id,
             },
         )
+        complete = response["Status"] in ("FAILED", "CREATE_COMPLETE")
+        if complete:
+            break
+        if sleep_time == max_sleep:
+            logger.debug(
+                "Still waiting on changeset for another %s seconds",
+                sleep_time
+            )
+        time.sleep(sleep_time)
+
+        # exponential backoff with max
+        sleep_time = min(sleep_time * 2, max_sleep)
+    if not complete:
+        raise exceptions.ChangesetDidNotStabilize(change_set_id)
+    return response
+
+
+def create_change_set(cfn_client, fqn, template_url, parameters, tags,
+                      replacements_only=False):
+    logger.debug("Attempting to create change set for stack: %s.", fqn)
+    response = retry_on_throttling(
+        cfn_client.create_change_set,
+        kwargs={
+            'StackName': fqn,
+            'TemplateURL': template_url,
+            'Parameters': parameters,
+            'Tags': tags,
+            'Capabilities': ["CAPABILITY_NAMED_IAM"],
+            'ChangeSetName': get_change_set_name(),
+        },
+    )
+    change_set_id = response["Id"]
+    response = wait_till_change_set_complete(
+        cfn_client, change_set_id
+    )
+    status = response["Status"]
+    if status == "FAILED":
+        status_reason = response["StatusReason"]
+        if "didn't contain changes" in response["StatusReason"]:
+            logger.debug(
+                "Stack %s did not change, not updating.",
+                fqn,
+            )
+            raise exceptions.StackDidNotChange
+        raise exceptions.UnhandledChangeSetStatus(
+            fqn, change_set_id, status, status_reason
+        )
+
+    execution_status = response["ExecutionStatus"]
+    if execution_status != "AVAILABLE":
+        raise exceptions.UnableToExecuteChangeSet(fqn,
+                                                  change_set_id,
+                                                  execution_status)
+
+    changes = response["Changes"]
+    return changes, change_set_id
+
+
+class Provider(AWSProvider):
+    """AWS Cloudformation Change Set Provider"""
+
+    def __init__(self, region, replacements_only=False, *args, **kwargs):
+        self.replacements_only = replacements_only
+        super(Provider, self).__init__(region=region, *args, **kwargs)
+
+    def update_stack(self, fqn, template_url, parameters, tags, diff=False,
+                     **kwargs):
+        changes, change_set_id = create_change_set(self.cloudformation, fqn,
+                                                   template_url, parameters,
+                                                   tags, **kwargs)
+
+        action = "replacements" if self.replacements_only else "changes"
+        full_changeset = changes
+        if self.replacements_only:
+            changes = requires_replacement(changes)
+
+        if changes:
+            output_summary(fqn, action, changes,
+                           replacements_only=self.replacements_only)
+            if not diff:
+                ask_for_approval(
+                    full_changeset=full_changeset,
+                    include_verbose=True,
+                )
+
+        if not diff:
+            retry_on_throttling(
+                self.cloudformation.execute_change_set,
+                kwargs={
+                    'ChangeSetName': change_set_id,
+                },
+            )
+        else:
+            retry_on_throttling(
+                self.cloudformation.delete_change_set,
+                kwargs={
+                    'ChangeSetName': change_set_id,
+                },
+            )
+
         return True